+++ src/skins/o2.tsx
@@ -223,7 +223,7 @@ export const getO2Skin: GetSkin = () => {
             brand: palette.o2BluePrimaryLight50,
             brandDark: palette.darkModeGrey6,
             inverse: palette.grey2,
-            neutralHigh: palette.grey4,
+            neutralHigh: palette.grey2,
             neutralMedium: palette.grey5,
             neutralLow: palette.darkModeGrey6,
             textPrimary: palette.grey2,